+++ src/aks-preview/HISTORY.rst
@@ -14,6 +14,7 @@ Pending
 * Implicitly enable istio when ingress or egress gateway is enabled for Azure Service Mesh.
 * Add `az aks nodepool delete-machines` command.
 * Update `az aks approuting zone` command to support private dns zones.
+* Vendor new SDK and bump API version to 2024-01-02-preview.
 
 1.0.0b6
 +++++++
@@ -22,6 +23,8 @@ Pending
 * Add `--ssh-access` to the `az aks update` command.
 * Add `--ssh-access` to the `az aks nodepool add` command.
 * Add `--ssh-access` to the `az aks nodepool update` command.
+* Implicitly enable istio when ingress or egress gateway is enabled for Azure Service Mesh.
+* Add `az aks nodepool delete-machines` command.
 
 1.0.0b5
 +++++++