+++ src/aks-preview/HISTORY.rst
@@ -11,6 +11,9 @@ To release a new version, please select a new version number (usually plus 1 to 
 
 Pending
 +++++++
+* Implicitly enable istio when ingress or egress gateway is enabled for Azure Service Mesh.
+* Add `az aks nodepool delete-machines` command.
+* Update `az aks approuting zone` command to support private dns zones.
 * Vendor new SDK and bump API version to 2024-01-02-preview.
 
 1.0.0b6